updaating ARP functions to be available on all configs
