+++ source/FreeRTOS_IPv4.c
@@ -42,10 +42,6 @@
 #include "FreeRTOS_IP.h"
 #include "FreeRTOS_IPv4.h"
 
-/* IPv4 multi-cast addresses range from 224.0.0.0.0 to 240.0.0.0. */
-#define ipFIRST_MULTI_CAST_IPv4    0xE0000000U          /**< Lower bound of the IPv4 multicast address. */
-#define ipLAST_MULTI_CAST_IPv4     0xF0000000U          /**< Higher bound of the IPv4 multicast address. */
-
 /* Just make sure the contents doesn't get compiled if IPv4 is not enabled. */
 /* *INDENT-OFF* */
     #if( ipconfigUSE_IPv4 != 0 )
@@ -183,31 +179,6 @@
     }
 
 #endif /* ( ipconfigDRIVER_INCLUDED_RX_IP_CHECKSUM == 1 ) */
-/*-----------------------------------------------------------*/
-
-/**
- * @brief Is the IP address an IPv4 multicast address.
- *
- * @param[in] ulIPAddress The IP address being checked.
- *
- * @return pdTRUE if the IP address is a multicast address or else, pdFALSE.
- */
-BaseType_t xIsIPv4Multicast( uint32_t ulIPAddress )
-{
-    BaseType_t xReturn;
-    uint32_t ulIP = FreeRTOS_ntohl( ulIPAddress );
-
-    if( ( ulIP >= ipFIRST_MULTI_CAST_IPv4 ) && ( ulIP < ipLAST_MULTI_CAST_IPv4 ) )
-    {
-        xReturn = pdTRUE;
-    }
-    else
-    {
-        xReturn = pdFALSE;
-    }
-
-    return xReturn;
-}
 /*-----------------------------------------------------------*/
 
 /**
