+++ source/FreeRTOS_IPv4_Utils.c
@@ -47,26 +47,6 @@
 
 /*-----------------------------------------------------------*/
 
-/**
- * @brief Set multicast MAC address.
- *
- * @param[in] ulIPAddress IP address.
- * @param[out] pxMACAddress Pointer to MAC address.
- */
-void vSetMultiCastIPv4MacAddress( uint32_t ulIPAddress,
-                                  MACAddress_t * pxMACAddress )
-{
-    uint32_t ulIP = FreeRTOS_ntohl( ulIPAddress );
-
-    pxMACAddress->ucBytes[ 0 ] = ( uint8_t ) 0x01U;
-    pxMACAddress->ucBytes[ 1 ] = ( uint8_t ) 0x00U;
-    pxMACAddress->ucBytes[ 2 ] = ( uint8_t ) 0x5EU;
-    pxMACAddress->ucBytes[ 3 ] = ( uint8_t ) ( ( ulIP >> 16 ) & 0x7fU ); /* Use 7 bits. */
-    pxMACAddress->ucBytes[ 4 ] = ( uint8_t ) ( ( ulIP >> 8 ) & 0xffU );  /* Use 8 bits. */
-    pxMACAddress->ucBytes[ 5 ] = ( uint8_t ) ( ( ulIP ) & 0xffU );       /* Use 8 bits. */
-}
-/*-----------------------------------------------------------*/
-
 /** @brief Do the first IPv4 length checks at the IP-header level.
  * @param[in] pucEthernetBuffer The buffer containing the packet.
  * @param[in] uxBufferLength The number of bytes to be sent or received.
