+++ source/include/FreeRTOS_IPv4_Utils.h
@@ -49,10 +49,6 @@
 /* FreeRTOS+TCP includes. */
 #include "FreeRTOS_IP.h"
 
-/* Set the MAC-address that belongs to a given IPv4 multi-cast address. */
-void vSetMultiCastIPv4MacAddress( uint32_t ulIPAddress,
-                                  MACAddress_t * pxMACAddress );
-
 /* The first IPv4 length checks at the IP-header level. */
 BaseType_t prvChecksumIPv4Checks( uint8_t * pucEthernetBuffer,
                                   size_t uxBufferLength,
