+++ test/unit-test/FreeRTOS_ARP/FreeRTOS_ARP_utest.c
@@ -9,6 +9,7 @@
 #include "mock_FreeRTOS_IP.h"
 #include "mock_FreeRTOS_IPv4.h"
 #include "mock_FreeRTOS_Routing.h"
+#include "mock_FreeRTOS_IP_Utils.h"
 #include "mock_FreeRTOS_IPv4_Utils.h"
 #include "mock_FreeRTOS_IP_Timers.h"
 #include "mock_FreeRTOS_IP_Private.h"
