+++ test/unit-test/FreeRTOS_IP_DiffConfig/FreeRTOS_IP_DiffConfig_utest.c
@@ -344,6 +344,7 @@ void test_prvAllowIPPacketIPv4_IncorrectSizeFields( void )
     pxIPHeader->ulSourceIPAddress = 0xC0C00101;
 
     FreeRTOS_FindEndPointOnIP_IPv4_ExpectAnyArgsAndReturn( NULL );
+    xIsIPv4Multicast_ExpectAnyArgsAndReturn( pdFALSE );
 
     eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
 
@@ -382,6 +383,7 @@ void test_prvAllowIPPacketIPv4_UDPCheckSumZero( void )
     pxIPHeader->ulSourceIPAddress = 0xC0C00101;
 
     FreeRTOS_FindEndPointOnIP_IPv4_ExpectAnyArgsAndReturn( NULL );
+    xIsIPv4Multicast_ExpectAnyArgsAndReturn( pdFALSE );
 
     eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
 
@@ -427,6 +429,7 @@ void test_prvAllowIPPacketIPv4_UDP_HappyPath( void )
     pxProtPack->xUDPPacket.xUDPHeader.usChecksum = 0xFF12;
 
     FreeRTOS_FindEndPointOnIP_IPv4_ExpectAnyArgsAndReturn( NULL );
+    xIsIPv4Multicast_ExpectAnyArgsAndReturn( pdFALSE );
 
     eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
 
@@ -466,6 +469,7 @@ void test_prvAllowIPPacketIPv4_TCP_HappyPath( void )
     pxIPHeader->ulSourceIPAddress = 0xC0C00101;
 
     FreeRTOS_FindEndPointOnIP_IPv4_ExpectAnyArgsAndReturn( NULL );
+    xIsIPv4Multicast_ExpectAnyArgsAndReturn( pdFALSE );
 
     eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
 
@@ -503,6 +507,7 @@ void test_prvProcessIPPacket_( void )
     pxIPHeader->ulSourceIPAddress = 0xC0C00101;
 
     FreeRTOS_FindEndPointOnIP_IPv4_ExpectAnyArgsAndReturn( NULL );
+    xIsIPv4Multicast_ExpectAnyArgsAndReturn( pdFALSE );
 
     eResult = prvProcessIPPacket( pxIPPacket, pxNetworkBuffer );
